pyproject.toml
[tool.poetry]
name = "epstein-files"
version = "1.9.3"
description = "Tools for working with the Jeffrey Epstein documents released in November 2025."
authors = ["Michel de Cryptadamus"]
readme = "README.md"
license = "GPL-3.0-or-later"

homepage = "https://michelcrypt4d4mus.github.io/epstein_text_messages/"
repository = "https://github.com/michelcrypt4d4mus/epstein_text_messages"

keywords = [
    "Epstein",
    "Jeffrey Epstein",
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]


[tool.poetry.dependencies]
python = "^3.11"
datefinder = "^0.7.3"
inflection = "^0.5.1"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.2.1"
requests = "^2.32.5"
rich = "^14.2.0"
rich-argparse-plus = "^0.3.1.4"
cairosvg = "^2.8.2"
pdfalyzer = {extras = ["extract"], version = "^1.19.6"}


# pdfalyzer = {path = "../../pdfalyzer", develop = true, extras=["extract"]}
# yaralyzer = {path = "../../yaralyzer", develop = true}


[tool.poetry.group.dev.dependencies]
bpython = "^0.26"
pytest = "^9.0.1"


[tool.poetry.scripts]
epstein_diff = 'epstein_files:epstein_diff'
epstein_generate = 'epstein_files:epstein_generate'
epstein_grep = 'epstein_files:epstein_grep'
epstein_pdf_path = 'epstein_files:epstein_pdf_path'
epstein_show = 'epstein_files:epstein_show'


[tool.poetry.urls]
Emails = "https://michelcrypt4d4mus.github.io/epstein_text_messages/all_emails_epstein_files_nov_2025.html"
Metadata = "https://michelcrypt4d4mus.github.io/epstein_text_messages/file_metadata_epstein_files_nov_2025.json"
TextMessages = "https://michelcrypt4d4mus.github.io/epstein_text_messages"
WordCounts = "https://michelcrypt4d4mus.github.io/epstein_text_messages/communication_word_count_epstein_files_nov_2025.html"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = ["--import-mode=importlib", "-v"]
pythonpath = ["."]
